+++ database/factories/IALFactory.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace Database\Factories;
+
+use App\Models\IAL;
+use App\Models\User;
+use Exception;
+use Illuminate\Database\Eloquent\Factories\Factory;
+
+class IALFactory extends Factory
+{
+	/**
+	 * The name of the factory's corresponding model.
+	 *
+	 * @var string
+	 */
+	protected $model = IAL::class;
+
+	/**
+	 * Define the model's default state.
+	 *
+	 * @return array
+	 * @throws Exception
+	 */
+	public function definition(): array
+	{
+		return [
+			'added'         => now(),
+			'user'          => User::class,
+			'type'          => 'log',
+			'wiki'          => 'wiki',
+			'comments'      => null,
+			'dpa'           => null,
+			'investigation' => null
+		];
+	}
+}