+++ database/migrations/2022_11_08_000001_update_dpas_primary_key.php
@@ -1,0 +1,32 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class UpdateDPASPrimaryKey extends Migration
+{
+	/**
+	 * Run the migrations.
+	 *
+	 * @return void
+	 */
+	public function up()
+	{
+		Schema::table( 'dpas', function( Blueprint $table ) {
+			$table->primary( 'id' );
+		} );
+	}
+
+	/**
+	 * Reverse the migrations.
+	 *
+	 * @return void
+	 */
+	public function down()
+	{
+		Schema::table( 'dpas', function( Blueprint $table ) {
+			$table->dropPrimary( 'id' );
+		} );
+	}
+}