+++ routes/web.php
@@ -2,12 +2,14 @@
 
 use App\Http\Controllers\AppealController;
 use App\Http\Controllers\DPAController;
+use App\Http\Controllers\IALController;
 use App\Http\Controllers\InvestigationController;
 use App\Http\Controllers\LoginController;
 use App\Http\Controllers\ReportController;
 use App\Http\Controllers\UserController;
 use App\Models\Appeal;
 use App\Models\DPA;
+use App\Models\IAL;
 use App\Models\Investigation;
 use App\Models\Report;
 use App\Models\User;
@@ -65,3 +67,9 @@ Route::redirect( '/gdpr', '/dpa' );
 Route::get( '/appeals', [ AppealController::class, 'index' ] )->middleware( 'auth' )->can( 'viewAny', Appeal::class );
 Route::get( '/appeal/{appeal}', [ AppealController::class, 'show' ] )->whereNumber( 'id' )->middleware( 'auth' )->can( 'view', 'appeal' );
 Route::patch( '/appeal/{appeal}', [ AppealController::class, 'update' ] )->whereNumber( 'id' )->middleware( 'auth' )->can( 'update', 'appeal' );
+
+/*
+ * Internal Actions Log web group
+ */
+Route::get( '/ial', [ IALController::class, 'index' ] )->middleware( 'auth' )->can( 'viewAny', IAL::class );
+Route::patch( '/ial/{ial}', [ IALController::class, 'update' ] )->whereNumber( 'id' )->middleware( 'auth' )->can( 'update', 'ial' );